# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Java Link Bridge API
  version: v0.1.5.24
servers:
  - url: http://localhost:8080
    description: local server
paths:
  /api/property/{id}/set:
    post:
      tags:
        - property-controller
      operationId: setProperty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPropertyRequest'
        required: true
      responses:
        '200':
          description: OK
  /api/instanceType/create:
    post:
      tags:
        - instance-type-controller
      operationId: createInstanceType
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstanceTypeRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InstanceTypeDTO'
  /api/instance/create:
    post:
      tags:
        - instance-controller
      operationId: createInstance
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstanceRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InstanceTypeDTO'
  /api/propertyType/getTypes/{instanceTypeId}:
    get:
      tags:
        - property-type-controller
      operationId: getPropertyTypes
      parameters:
        - name: instanceTypeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PropertyTypeDTO'
  /api/property/getProperties/{instanceId}:
    get:
      tags:
        - property-controller
      operationId: getProperties
      parameters:
        - name: instanceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PropertyDTO'
  /api/property/getInstances/{typeId}:
    get:
      tags:
        - instance-controller
      operationId: getInstances
      parameters:
        - name: typeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstanceTypeDTO'
  /api/instanceType/getTypes:
    get:
      tags:
        - instance-type-controller
      operationId: getInstanceTypes
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstanceTypeDTO'
  /api/instance/getAll:
    get:
      tags:
        - instance-controller
      operationId: getAllInstances
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstanceTypeDTO'
  /api/instance/get/{id}:
    get:
      tags:
        - instance-controller
      operationId: getInstance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InstanceTypeDTO'
  /api/instanceType/delete/{id}:
    delete:
      tags:
        - instance-type-controller
      operationId: deleteInstanceType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  /api/instance/delete/{id}:
    delete:
      tags:
        - instance-controller
      operationId: deleteInstance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
components:
  schemas:
    SetPropertyRequest:
      type: object
      properties:
        propertyId:
          type: integer
          format: int64
        data:
          type: object
    CreateInstanceTypeRequest:
      type: object
      properties:
        parentInstanceId:
          type: integer
          format: int64
        name:
          type: string
    InstanceTypeDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    CreateInstanceRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    PropertyTypeDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    PropertyDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string